+++ packages/react-querybuilder/src/RuleGroup.tsx
@@ -1,4 +1,4 @@
-import type { RuleGroupProps } from '@react-querybuilder/ts';
+import type { RuleGroupProps, RuleGroupType, RuleGroupTypeAny } from '@react-querybuilder/ts';
 import { clsx } from 'clsx';
 import type { MouseEvent as ReactMouseEvent } from 'react';
 import { Fragment, useMemo } from 'react';
@@ -26,6 +26,8 @@ export const RuleGroup = ({
   previewRef = null,
   dragRef = null,
   dropRef = null,
+  isDragging = false,
+  isOver = false,
 }: RuleGroupProps) => {
   const {
     classNames,
@@ -57,19 +59,27 @@ export const RuleGroup = ({
   const { onGroupAdd, onGroupRemove, onPropChange, onRuleAdd, moveRule } = actions;
   const disabled = !!parentDisabled || !!disabledProp;
 
-  const { rules, not } = ruleGroup ? ruleGroup : { rules: rulesProp, not: notProp };
+  const ruleGroupObject = ruleGroup
+    ? { ...ruleGroup }
+    : ({ rules: rulesProp, not: notProp } as RuleGroupTypeAny);
+  const { rules, not } = ruleGroupObject;
   let combinator: string = defaultCombinators[0].name;
   if (ruleGroup && 'combinator' in ruleGroup) {
     combinator = ruleGroup.combinator;
   } else if (!ruleGroup) {
     combinator = combinatorProp ?? combinator;
   }
+  if (!independentCombinators) {
+    (ruleGroupObject as RuleGroupType).combinator = combinator;
+  }
 
   useDeprecatedProps('ruleGroup', !!ruleGroup);
 
   const classNamesMemo = useMemo(
     () => ({
-      header: clsx(standardClassnames.header, classNames.header),
+      header: clsx(standardClassnames.header, classNames.header, {
+        [standardClassnames.dndOver]: isOver,
+      }),
       dragHandle: clsx(standardClassnames.dragHandle, classNames.dragHandle),
       combinators: clsx(standardClassnames.combinators, classNames.combinators),
       notToggle: clsx(standardClassnames.notToggle, classNames.notToggle),
@@ -80,7 +90,19 @@ export const RuleGroup = ({
       removeGroup: clsx(standardClassnames.removeGroup, classNames.removeGroup),
       body: clsx(standardClassnames.body, classNames.body),
     }),
-    [classNames]
+    [
+      classNames.addGroup,
+      classNames.addRule,
+      classNames.body,
+      classNames.cloneGroup,
+      classNames.combinators,
+      classNames.dragHandle,
+      classNames.header,
+      classNames.lockGroup,
+      classNames.notToggle,
+      classNames.removeGroup,
+      isOver,
+    ]
   );
 
   useReactDndWarning(
@@ -159,7 +181,10 @@ export const RuleGroup = ({
   const outerClassName = clsx(
     standardClassnames.ruleGroup,
     classNames.ruleGroup,
-    { [standardClassnames.disabled]: disabled },
+    {
+      [standardClassnames.disabled]: disabled,
+      [standardClassnames.dndDragging]: isDragging,
+    },
     validationClassName
   );
 
@@ -231,6 +256,7 @@ export const RuleGroup = ({
           disabled={disabled}
           context={context}
           validation={validationResult}
+          ruleOrGroup={ruleGroupObject}
         />
         <AddGroupActionControlElement
           testID={TestID.addGroup}
@@ -244,6 +270,7 @@ export const RuleGroup = ({
           disabled={disabled}
           context={context}
           validation={validationResult}
+          ruleOrGroup={ruleGroupObject}
         />
         {showCloneButtons && path.length >= 1 && (
           <CloneGroupActionControlElement
@@ -258,6 +285,7 @@ export const RuleGroup = ({
             disabled={disabled}
             context={context}
             validation={validationResult}
+            ruleOrGroup={ruleGroupObject}
           />
         )}
         {showLockButtons && (
@@ -274,6 +302,7 @@ export const RuleGroup = ({
             disabledTranslation={parentDisabled ? undefined : translations.lockGroupDisabled}
             context={context}
             validation={validationResult}
+            ruleOrGroup={ruleGroupObject}
           />
         )}
         {path.length >= 1 && (
@@ -289,6 +318,7 @@ export const RuleGroup = ({
             disabled={disabled}
             context={context}
             validation={validationResult}
+            ruleOrGroup={ruleGroupObject}
           />
         )}
       </div>